Fix Github Font Preview mini glyph fill color (#14)
The font preview userscript was previously drawing the glyphs within the
preview pages without specifying the fill color as was done when drawing
the large glyph. This worked fine when the default fill color was the
standard black, but when the color was very light, like it is in dark
themes such as Github Dark, this resulted in having nearly-white glyphs
drawn on top of a nearly-white background. The readability problem this
caused has been fixed by drawing all the glyphs, both the small preview
glyphs and the large detail glyph, using a specific color. That color is
controlled by the "glyphFillColor" variable, which is simply a rename of
the "bigGlyphFillColor" variable but still using the same default color.

You can see a before/after of the UI changes involved in the fix here:
https://imgur.com/1K7Tb7j

Signed-off-by: Emily Mabrey <emilymabrey93@gmail.com>